# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AiMessageRole = (typeof AiMessageRole)[keyof typeof AiMessageRole];

export const AiMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface AiMessage {
  role: AiMessageRole;
  content: string;
}

export interface AiChatRequest {
  system: string;
  messages: AiMessage[];
}

export interface AiChatResponse {
  text: string;
}

export interface AiError {
  error: string;
}