# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Send a message to an AI persona
 */
export const AiChatBody = zod.object({
  system: zod.string(),
  messages: zod.array(
    zod.object({
      role: zod.enum(["user", "assistant"]),
      content: zod.string(),
    }),
  ),
});

export const AiChatResponse = zod.object({
  text: zod.string(),
});