# lib/api-zod/src/generated/types/aiChatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiMessage } from "./aiMessage";

export interface AiChatRequest {
  system: string;
  messages: AiMessage[];
}